Added a Raspcontrol Update button
This shows an update button on the menu bar along the top and sources
the update from the .git.
If git-core isn't installed then it does this before attempting a
download.

Due to rights access I had to download it to a temp directory
(/var/www/RaspcontrolUpdate) before renaming to the original
/var/www/Raspcontrol.

Tested multiple times to the best of my ability and it works as expected
on my machine, however due to me being very much a Linux beginner the
permissions on /www/ may differ to that of a real user so I'd appreciate
some double checking before merging. I don't forsee any issues, but as I
say, my setup may be different than a normal user.

I'm also learning to use github so sorry if these updates reach you in a
way that don't make any sense...I still have more reading to do about
branches/commits etc etc.

I will be looking into making it so the button only appears when there's
a valid update present, there's no point allowing users to use bandwidth
to update Raspcontrol with an existing version. I'll look into this as a
future update.