Using a gunction to translate bytes into Mb/Gb
Using the suggestion of a function to run the repeated actions of
translating bytes into MB/GB/TB as posted here
(https://github.com/Bioshox/Raspcontrol/issues/67)

I am unsure what the extent of the changes were in regards to the
calculations so I have kept the existing logic as provided by you and
just put it into a function. If the calculation is to be changed in the
future then only the function needs to be changed rather than 2 areas of
text.