Auto detects the install path for Raspcontrol
No user input required when running an update. Variable is set at login
time and is used as the basis for the upgrade.

Tested on my machine and as it's dynamic it should work for everyone. On
the downside it's leaving you tied into apache more and more which is
something you were considering trying to get away from.